secondary project/BACKEND/listing/serializers.py
from rest_framework import serializers
from .models import DoctorProfile

class DoctorProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = DoctorProfile
        fields = (
            'id',
            'practice_name',
            'rating',
            'image_url'
            )
